backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai
debugpy==1.8.0